changed gallery to new gallery
